requirement.txt
streamlit
pandas
matplotlib
seaborn
numpy
plotly
scikit-learn
statsmodels
fpdf
shap
tensorflow
python-pptx
xlsxwriter

